+++ src/Components/MyItems/MyItems.js
@@ -1,42 +1,66 @@
+import { signOut } from 'firebase/auth';
 import React, { useEffect, useState } from 'react';
 import { useAuthState } from 'react-firebase-hooks/auth';
+import { useNavigate } from 'react-router-dom';
 import auth from '../../firebase.init';
+import Footer from '../Footer/Footer';
 import MyItemList from '../MyItemList/MyItemList';
 
 const MyItems = () => {
+    const navigate = useNavigate();
     const [myItems, setMyitems] = useState([]);
     const [user] = useAuthState(auth);
 
 
     useEffect(() => {
         const email = user?.email;
-        console.log(email)
+
         const url = `http://localhost:5000/myitem?email=${email}`;
-        fetch(url)
-            .then((response) => response.json())
-            .then((data) => setMyitems(data));
-    }, [user.email]);
+
+        try {
+            fetch(url, {
+                headers: {
+                    authrization: `Bearer ${localStorage.getItem('accessToken')}`
+                }
+            })
+                .then((response) => response.json())
+                .then((data) => setMyitems(data))
+        }
+        catch (error) {
+            console.log(error.message)
+            if (error.response.status === 401 || error.response.status === 403) {
+                signOut(auth);
+                navigate('/login')
+            }
+        }
+
+    }, [user?.email, navigate]);
 
     const handleDelete = DeleteId => {
         console.log(DeleteId)
-        const url = `http://localhost:5000/item/${DeleteId}`
-        fetch(url, {
-            method: 'DELETE',
-        })
-            .then((response) => response.json())
-            .then((data) => {
-                if (data.deletedCount > 0) {
-                    const remains = myItems.filter(item => item._id !== DeleteId)
-                    setMyitems(remains)
-                }
+        const proceed = window.confirm('Are you sure?')
+        if (proceed) {
+            const url = `http://localhost:5000/item/${DeleteId}`
+            fetch(url, {
+                method: 'DELETE',
+            })
+                .then((response) => response.json())
+                .then((data) => {
+                    if (data?.deletedCount > 0) {
+                        const remains = myItems.filter(item => item._id !== DeleteId)
+                        setMyitems(remains)
+                    }
 
-            });
+                });
+        }
+
+
 
     }
 
 
     return (
-        <div className='h-screen bg-[#F0ECE3]'>
+        <div className=' bg-[#F0ECE3]'>
             <div className='w-11/12 mx-auto py-12'>
                 <h2 className='text-center font-condensed font-bold text-4xl pb-8'>My Items : {myItems.length}</h2>
                 <div className="overflow-x-auto relative">
@@ -67,7 +91,6 @@ const MyItems = () => {
                             </tr>
                         </thead>
                         <tbody>
-
                             {
                                 myItems.map(myItem =>
                                     <MyItemList
@@ -82,6 +105,7 @@ const MyItems = () => {
                     </table>
                 </div>
             </div>
+            <Footer />
         </div>
     );
 };
